Add tests for custom transforms
